test: validation: rename time test header file
Header file name time.h is ambiguos since C has a library header
file with the same name.

Reviewed-and-tested-by: Bill Fischofer <bill.fischofer@linaro.org>
