+++ test/common_plat/validation/api/time/Makefile.am
@@ -7,4 +7,4 @@ test_PROGRAMS = time_main$(EXEEXT)
 dist_time_main_SOURCES = time_main.c
 time_main_LDADD = libtesttime.la $(LIBCUNIT_COMMON) $(LIBODP)
 
-EXTRA_DIST = time.h
+EXTRA_DIST = time_test.h
